# Optimize `BeEquivalentTo`