In #1932 I noticed that we generate the equivalency summary even when a test does not fail.
That didn't look right to me.

### Benchmark 1: Comparing two simple objects
<details>

```cs
[MemoryDiagnoser]
[SimpleJob(RuntimeMoniker.Net472)]
[SimpleJob(RuntimeMoniker.Net60)]
public class BeEquivalentToBenchmarks
{
    public class MyClass
    {
        public int MyProperty { get; set; }
    }

    private readonly MyClass a = new();
    private readonly MyClass b = new();

    [Benchmark]
    public AndConstraint<ObjectAssertions> BeEquivalentTo() => a.Should().BeEquivalentTo(b);
}
```
</details>

develop
|         Method |              Runtime |      Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |----------:|----------:|----------:|-------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 |  8.888 us | 0.1760 us | 0.2524 us | 2.0447 | 0.0153 |     13 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 11.181 us | 0.1033 us | 0.0915 us | 2.1820 | 0.0153 |     13 KB |


373a6add033fec1baf7870fcd10ba4ce61be2e44
|         Method |              Runtime |      Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |----------:|----------:|----------:|-------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 |  7.830 us | 0.0923 us | 0.0863 us | 1.7090 | 0.0153 |     11 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 10.427 us | 0.1348 us | 0.1195 us | 1.8463 | 0.0153 |     11 KB |


6c5db114828617e2047588bea7750705d18e3893
|         Method |              Runtime |      Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |----------:|----------:|----------:|-------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 |  6.972 us | 0.0881 us | 0.0824 us | 1.5564 | 0.0153 |     10 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 10.023 us | 0.1090 us | 0.1020 us | 1.6785 | 0.0153 |     10 KB |


72c2e256ce1d8e630e049f410d5aa02f8af56abd
|         Method |              Runtime |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |---------:|----------:|----------:|-------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 | 7.239 us | 0.0411 us | 0.0385 us | 1.4954 | 0.0153 |      9 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 9.646 us | 0.0978 us | 0.0915 us | 1.6785 | 0.0153 |     10 KB |


a50e07a0da331f4e047205a54cd63e9100fa1ce6
|         Method |              Runtime |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |---------:|----------:|----------:|-------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 | 6.732 us | 0.0667 us | 0.0624 us | 1.3809 | 0.0153 |      8 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 9.559 us | 0.1534 us | 0.1435 us | 1.5717 | 0.0153 |     10 KB |


8eeb20c6744eea033ed732d3125ce37501045433
|         Method |              Runtime |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |---------:|----------:|----------:|-------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 | 6.833 us | 0.0264 us | 0.0247 us | 1.3199 | 0.0153 |      8 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 9.518 us | 0.1333 us | 0.1247 us | 1.5106 | 0.0153 |      9 KB |

### Benchmark 2: Comparing lists of complex objects
<details>

`BeEquivalentToWithDeeplyNestedStructures` with these modifications

```cs
[Params(5)]
public int N { get; set; }

[Params(4)]
public int Depth { get; set; }

[GlobalSetup]
public void GlobalSetup()
{
	subject = Enumerable.Range(0, N).Select(_ => CreateComplex(Depth)).ToList();
	expectation = Enumerable.Range(0, N).Select(_ => CreateComplex(Depth)).ToList();

	AssertionOptions.AssertEquivalencyUsing(e => e.WithoutStrictOrdering());
}
```
</details>

develop
|         Method |              Runtime | N | Depth |     Mean |   Error |  StdDev |   Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |-- |------ |---------:|--------:|--------:|--------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 | 5 |     4 | 243.9 us | 1.60 us | 1.49 us | 44.9219 | 1.9531 |    278 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 5 |     4 | 345.5 us | 3.87 us | 3.62 us | 49.3164 | 1.9531 |    303 KB |


8eeb20c6744eea033ed732d3125ce37501045433 - `RequiresConversion`
|         Method |              Runtime | N | Depth |     Mean |   Error |  StdDev |   Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |-- |------ |---------:|--------:|--------:|--------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 | 5 |     4 | 222.2 us | 1.87 us | 1.46 us | 34.6680 | 0.9766 |    215 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 5 |     4 | 323.7 us | 5.21 us | 4.88 us | 40.0391 | 1.4648 |    247 KB |


55c79b340bc85f87c144351804cf9e6ae8975f61 `Node.GetHashCode`
|         Method |              Runtime | N | Depth |     Mean |   Error |  StdDev |   Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |-- |------ |---------:|--------:|--------:|--------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 | 5 |     4 | 216.6 us | 1.21 us | 1.08 us | 33.2031 | 1.2207 |    205 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 5 |     4 | 305.7 us | 3.23 us | 3.03 us | 38.0859 | 1.4648 |    235 KB |

55c79b340bc85f87c144351804cf9e6ae8975f61 `Node.Equals`
|         Method |              Runtime | N | Depth |     Mean |   Error |  StdDev |   Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |-- |------ |---------:|--------:|--------:|--------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 | 5 |     4 | 216.5 us | 1.17 us | 1.10 us | 32.7148 | 1.2207 |    201 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 5 |     4 | 306.4 us | 4.11 us | 3.64 us | 37.1094 | 0.9766 |    231 KB |


55c79b340bc85f87c144351804cf9e6ae8975f61 Cache `PathName`
|         Method |              Runtime | N | Depth |     Mean |   Error |  StdDev |   Gen 0 |  Gen 1 | Allocated |
|--------------- |--------------------- |-- |------ |---------:|--------:|--------:|--------:|-------:|----------:|
| BeEquivalentTo |             .NET 6.0 | 5 |     4 | 209.4 us | 0.99 us | 0.92 us | 31.2500 | 1.2207 |    192 KB |
| BeEquivalentTo | .NET Framework 4.7.2 | 5 |     4 | 288.2 us | 2.05 us | 1.92 us | 35.6445 | 0.9766 |    221 KB |